Move capitalizeWords func to own file
